nlp-service/requirements.txt
fastapi
uvicorn
scikit-learn
python-multipart
PyPDF2
python-docx
numpy
